If at first you don't succeed... Review URL: https://chromiumcodereview.appspot.com//9117051
git-svn-id: https://dart.googlecode.com/svn/branches/bleeding_edge/dart@3574 260f80e4-7a28-3924-810f-c04153c831b5

+++ runtime/vm/debugger_api_impl_test.cc
@@ -198,12 +198,12 @@ TEST_CASE(Debug_LookupSourceLine) {
   intptr_t num_libs;
   Dart_ListLength(libs, &num_libs);
   EXPECT(num_libs > 0);
-  for (intptr_t i = 0; i < num_libs; i++) {
+  for (int i = 0; i < num_libs; i++) {
     Dart_Handle lib_url = Dart_ListGetAt(libs, i);
     EXPECT(Dart_IsString(lib_url));
     char const* chars;
     Dart_StringToCString(lib_url, &chars);
-    printf("Lib %ld: %s\n", i, chars);
+    printf("Lib %d: %s\n", i, chars);
 
     Dart_Handle scripts = Dart_GetScriptURLs(lib_url);
     EXPECT(Dart_IsList(scripts));
@@ -214,7 +214,7 @@ TEST_CASE(Debug_LookupSourceLine) {
       Dart_Handle script_url = Dart_ListGetAt(scripts, i);
       char const* chars;
       Dart_StringToCString(script_url, &chars);
-      printf("  script %ld: '%s'\n", i + 1, chars);
+      printf("  script %d: '%s'\n", i + 1, chars);
     }
   }
